requirements.txt


streamlit
langchain-community
huggingface_hub
python-dotenv
